Rename release/utils/github-client.ts to pull-requests.ts
